pyproject.toml
[project]
name = "aimlinternassignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "textblob>=0.19.0",
]
